feat: 支持安装 release pods

+++ lib/lg_pod_plugin/gitlab_download.rb
@@ -23,54 +23,6 @@ module LgPodPlugin
       self.branch = options[:branch]
       self.commit = options[:commit]
     end
-    # 封装 git clone命令
-    def git_download_command(temp_name, git, branch, tag)
-      cmds = ['git']
-      cmds << "clone"
-      cmds << "#{git}"
-      cmds << "#{temp_name} "
-      cmds << "--template="
-      cmds << "--single-branch --depth 1"
-      if branch
-        cmds << "--branch"
-        cmds << branch
-      elsif tag
-        cmds << "--branch"
-        cmds << tag
-      end
-      cmds_to_s = cmds.join(" ")
-      LgPodPlugin.log_blue cmds_to_s
-      system(cmds_to_s)
-    end
-
-    def git_clone_by_branch(path, temp_name)
-      download_temp_path = path.join(temp_name)
-      if self.git && self.branch
-        git_download_command(temp_name, self.git, self.branch, nil)
-      else
-        git_download_command(temp_name, self.git, nil, nil)
-        if File.exist?(temp_name)
-          system("git -C #{download_temp_path.to_path} rev-parse HEAD")
-        end
-      end
-      download_temp_path
-    end
-
-    def git_clone_by_tag(path, temp_name)
-      git_download_command(temp_name, self.git, nil, self.tag)
-      path.join(temp_name)
-    end
-
-    # git clone commit
-    def git_clone_by_commit(path, temp_name)
-      Git.init(temp_name)
-      FileUtils.chdir(temp_name)
-      LgPodPlugin.log_blue "git clone #{self.git}"
-      system("git remote add origin #{self.git}")
-      system("git fetch origin #{self.commit}")
-      system("git reset --hard FETCH_HEAD")
-      path.join(temp_name)
-    end
 
     # clone 代码仓库
     def git_clone_repository(path)
@@ -84,10 +36,10 @@ module LgPodPlugin
           elsif self.git.include?("https://github.com")
             return git_archive.github_download_tag_zip path, temp_name
           else
-            return self.git_clone_by_tag(path, temp_name)
+            return git_archive.git_clone_by_tag(path, temp_name)
           end
         rescue
-          return self.git_clone_by_tag(path, temp_name)
+          return git_archive.git_clone_by_tag(path, temp_name)
         end
       elsif self.git && self.branch
         begin
@@ -96,10 +48,10 @@ module LgPodPlugin
           elsif self.git.include?("https://github.com")
             return git_archive.github_download_branch_zip path, temp_name
           else
-            return self.git_clone_by_branch(path, temp_name)
+            return git_archive.git_clone_by_branch(path, temp_name)
           end
         rescue
-          return self.git_clone_by_branch(path, temp_name)
+          return git_archive.git_clone_by_branch(path, temp_name)
         end
       elsif self.git && self.commit
         if LUtils.is_use_gitlab_archive_file(self.git)
@@ -107,7 +59,7 @@ module LgPodPlugin
         elsif self.git.include?("https://github.com")
           return git_archive.github_download_commit_zip path, temp_name
         else
-          return self.git_clone_by_commit(path, temp_name)
+          return git_archive.git_clone_by_commit(path, temp_name)
         end
       elsif self.git
         if LUtils.is_use_gitlab_archive_file(self.git)
@@ -115,7 +67,7 @@ module LgPodPlugin
         elsif self.git.include?("https://github.com")
           return git_archive.github_download_branch_zip path, temp_name
         else
-          return self.git_clone_by_branch(path, temp_name)
+          return git_archive.git_clone_by_branch(path, temp_name)
         end
       end
 
@@ -124,21 +76,13 @@ module LgPodPlugin
     # git 预下载
     def pre_download_git_repository
       temp_path = LFileManager.download_director.join("temp")
-      if temp_path.exist?
-        FileUtils.rm_rf(temp_path)
-      end
+      FileUtils.rm_rf(temp_path) if temp_path.exist?
       lg_pod_path = LRequest.shared.cache.cache_root
-      unless lg_pod_path.exist?
-        lg_pod_path.mkdir(0700)
-      end
+      lg_pod_path.mkdir(0700) unless lg_pod_path.exist?
       get_temp_folder = git_clone_repository(lg_pod_path)
       #下载 git 仓库失败
-      unless get_temp_folder && get_temp_folder.exist?
-        return nil
-      end
-      if LRequest.shared.single_git
-        LgPodPlugin::LCache.cache_pod(self.name, get_temp_folder, { :git => self.git })
-      end
+      return nil unless (get_temp_folder && get_temp_folder.exist?)
+      LgPodPlugin::LCache.cache_pod(self.name, get_temp_folder, { :git => self.git }) if LRequest.shared.single_git
       LgPodPlugin::LCache.cache_pod(self.name, get_temp_folder, LRequest.shared.get_cache_key_params)
       FileUtils.chdir(LFileManager.download_director)
       FileUtils.rm_rf(lg_pod_path)
@@ -162,7 +106,7 @@ module LgPodPlugin
           new_commit = pod_info.commit if pod_info
           return [branch, new_commit]
         end
-        new_commit = LUtils.commit_from_ls_remote(result, branch)
+        new_commit, new_branch = LUtils.commit_from_ls_remote(result, branch)
         if new_commit
           LSqliteDb.shared.insert_pod_refs(name, git, branch, tag, new_commit)
         end
@@ -178,13 +122,14 @@ module LgPodPlugin
           id = LPodLatestRefs.get_pod_id(name, git, branch, tag)
           pod_info = LSqliteDb.shared.query_pod_refs(id)
           new_commit = pod_info.commit if pod_info
-          return [nil, new_commit]
+          new_branch = pod_info.branch if pod_info
+          return [new_branch, new_commit]
         end
-        new_commit = LUtils.commit_from_ls_remote(result, tag)
+        new_commit, new_branch = LUtils.commit_from_ls_remote(result, tag)
         if new_commit
           LSqliteDb.shared.insert_pod_refs(name, git, branch, tag, new_commit)
         end
-        return [nil, new_commit]
+        return [new_branch, new_commit]
       elsif commit
         return nil, commit
       else
@@ -198,13 +143,14 @@ module LgPodPlugin
           id = LPodLatestRefs.get_pod_id(name, git, branch, tag)
           pod_info = LSqliteDb.shared.query_pod_refs(id)
           new_commit = pod_info.commit if pod_info
-          return [nil, new_commit]
+          new_branch = pod_info.branch if pod_info
+          return [new_branch, new_commit]
         end
-        new_commit = LUtils.commit_from_ls_remote(result, "HEAD")
+        new_commit, new_branch = LUtils.commit_from_ls_remote(result, "HEAD")
         if new_commit
-          LSqliteDb.shared.insert_pod_refs(name, git, branch, tag, new_commit)
+          LSqliteDb.shared.insert_pod_refs(name, git, new_branch, tag, new_commit)
         end
-        return [nil, new_commit]
+        return [new_branch, new_commit]
       end
     end
